+++ CHANGELOG.md
@@ -11,7 +11,7 @@ These changes likely break some apps, please read the
 [upgrading doc](https://developer.fyne.io/api/v2.0/upgrading) for more info
 
 * The second parameter of `theme.NewThemedResource` was removed, it was previously ignored
-* `fyne.Theme` moved to `fyne.LegacyTheme`
+* `fyne.Theme` moved to `fyne.LegacyTheme` and can be load to a new theme using `theme.FromLegacy`
 * A new, more flexible, Theme interface has been created that we encourage developers to use
 
 ### Added